Delete spruce_trapdoor.png
I propose for the vanilla (default) texture of the spruce trapdoor to be returned across all servers. 
The spruce trapdoor is one of,  if not the most used trapdoor types for building. Its texture matches spruce type blocks perfectly and also goes well with some other wood types, for example dark oak. it also seamlessly connects when placed in a bridge like fasision because of its planky nature, which can not be said about the new texture, because it looks like a crate. The seamless nature of the said block and texture fitting with many wood related blocks has been established for a long time in the building community and changing it means disrupting the intentional design of many builds. Another point is the iron lining used in the vanilla texture. It gives a wood focused build a slightly enhanced feel of multiple materials being used. In sumarry the spruce trapdoor (with the vanilla texture) is a block very important for decorating builds which utilise wood as a material. This texture has been used for a long time and is loved by many players.